refactor(projects): рефакторинг проектов
Jira issue: VEGA-551
